requirements.txt
py7zr==0.20.8
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
